### 2023-07-13 #5606 @lucydot
I'll add that there are reviewer docs here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
