+++ _sources/index.rst.txt
@@ -72,18 +72,22 @@ Step 2: visualize tracked data
 
 ------------
 
-**Guides**
+**Content**
 
 .. toctree::
   :maxdepth: 2
+
+  Documentation Home <self>
+
+.. toctree::
+  :maxdepth: 2
+  :caption: Guides
 
   Guides <guides/index>
 
-
-**API references**
-
 .. toctree::
   :maxdepth: 2
+  :caption: API references
 
   API References <apidocs/index>
 
