+++ docs/requirements-doc.txt
@@ -22,4 +22,4 @@ ipywidgets==8.0.6
 matplotlib==3.7.1
 pandas==2.0.2
 pylatexenc==2.10
-qiskit-sphinx-theme~=1.13.1
+qiskit-sphinx-theme~=1.14.0rc1
