+++ .github/workflows/docs.yml
@@ -26,7 +26,7 @@ jobs:
         run: |
           tox -edocs
       - name: Deploy docs
-        if: startsWith(github.ref, 'refs/heads/stable/') || github.ref == 'refs/heads/main'
+        #if: startsWith(github.ref, 'refs/heads/stable/') || github.ref == 'refs/heads/main'
         uses: peaceiris/actions-gh-pages@v3
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
