+++ docs/conf.py
@@ -16,9 +16,9 @@ author = ""
 _rootdir = Path(__file__).parent.parent
 
 # The full version, including alpha/beta/rc tags
-release = "0.0.0"
+release = "0.2.0"
 # The short X.Y version
-version = "0.0"
+version = "0.2"
 
 extensions = [
     "sphinx.ext.napoleon",
