### 2023-07-18 6f96957
add transparent logo
