### 2023-07-20 #5606 @lucydot
Hi @IceKhan13,

I am going on Annual Leave until the 8th of August. I will ping a couple more possible reviewers now, so hopefully we can get a review going in the next couple of days.

@obliviateandsurrender @DevenPanchal @mkareshk - are you available to review this submission for JOSS?
I believe a couple of you are new reviewers to JOSS, so welcome! Our process is a little different to more traditional journals, being Github-based and completely transparent. There are more details on the reviewer docs: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
