+++ docs/conf.py
@@ -11,14 +11,14 @@ sys.path.append(os.path.abspath("../client"))
 
 project = "PurpleCaffeine"
 copyright = "2023"  # pylint: disable=redefined-builtin
-author = ""
+author = "Iskandar Sitdikov, Michaël Rollin, Ansah Mohammad, Luis Eduardo Martínez Hérnandez"
 
 _rootdir = Path(__file__).parent.parent
 
 # The full version, including alpha/beta/rc tags
-release = "0.0.0"
+release = "0.2.0"
 # The short X.Y version
-version = "0.0"
+version = "0.2"
 
 extensions = [
     "sphinx.ext.napoleon",
@@ -32,6 +32,7 @@ extensions = [
     "reno.sphinxext",
     "nbsphinx",
     "sphinx_copybutton",
+    "qiskit_sphinx_theme",
 ]
 templates_path = ["_templates"]
 numfig = True
@@ -42,8 +43,12 @@ add_module_names = False
 modindex_common_prefix = ["purple_caffeine_project."]
 
 # html theme options
-html_static_path = ["_static"]
-# html_logo = "_static/images/logo.png"
+html_theme = "qiskit-ecosystem"
+html_title = f"{project} {release}"
+html_logo = "images/transparent_logo.png"
+html_theme_options = {
+    "sidebar_hide_name": True
+}
 
 # autodoc/autosummary options
 autosummary_generate = True
