+++ docs/index.rst
@@ -1,10 +1,6 @@
-##############
-PurpleCaffeine
-##############
+.. image:: /images/readme_logo.png
 
 PurpleCaffeine is a software for tracking quantum programs and experiments.
-
-.. image:: /images/readme_logo.png
 
 The source code to the project is available `on GitHub <https://github.com/IceKhan13/purplecaffeine>`_.
 
@@ -72,19 +68,17 @@ Step 2: visualize tracked data
 
 ------------
 
-**Guides**
+**Content**
+
+.. toctree::
+  :maxdepth: 2
+
+  Documentation Home <self>
 
 .. toctree::
   :maxdepth: 2
 
   Guides <guides/index>
-
-
-**API references**
-
-.. toctree::
-  :maxdepth: 2
-
   API References <apidocs/index>
 
 .. Hiding - Indices and tables
