+++ docs/guides/index.rst
@@ -1,3 +1,7 @@
+#############
+How-To Guides
+#############
+
 .. nbgallery::
    :glob:
 
