### 2023-07-21 #5606 @lucydot
That's great @obliviateandsurrender and @DevenPanchal 🎈

@DevenPanchal  - not here (this is the _pre-_review), I will add you as reviewers and ask our editorial-bot to generate a review-thread. There will be reviewer instructions at the top of that thread.

The review process at JOSS is a little different in that it is more an on-going conversation between the authors, reviewers, editorialbot and editor. We encourage you to raise and discuss points as they arise rather than listing everything at the end in a single post. Apologies I will not be more hands on for the start of your review (I am on leave until the 8th of August). If questions do arise it may be that you want to hold back on progressing/completing the review until my return. You should be able to get started I think without too many hurdles: also there are reviewer guidelines are here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
