### 2023-07-21 #5673 @DevenPanchal
@IceKhan13 :  2 concerns before I finish the review:

1.  Committer https://github.com/Eric-Arellano who is a contributor to the repository, has not been mentioned in the paper.

I need this information to address the below point in the checklist:
_Contribution and authorship: Has the submitting author (@IceKhan13) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?_

2. The dependencies have been mentioned at the top of the repo. But I feel, the dependencies could be mentioned again just before the installation instructions. This would make it easier to read.

I need this information to address the below point in the checklist:
_Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution._
