### 2023-07-21 #5673 @IceKhan13
Hello @DevenPanchal

1. Eric responsible to unify docs across ecosystem of Qiskit tools. Therefore they out of list

> Contribution and authorship: Has the submitting author (@IceKhan13) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

answer to both questions is `yes`

2. Dependencies are listed in `client` directory requirements file [here](https://github.com/IceKhan13/purplecaffeine/blob/main/client/requirements.txt). And installed automatically when user run `pip install purplecaffeine` which is configured [here](https://github.com/IceKhan13/purplecaffeine/blob/main/client/setup.py#L9).
