### 2023-07-21 #5606 @WhiteSymmetry
> Hi @IceKhan13,
>
> Thanks for the suggestions.
>
> @Hanrui-Wang, @TejasAvinashShetty, @WhiteSymmetry - are you available to review this submission for JOSS? I believe a couple of you are new reviewers to JOSS, so welcome! Our process is a little different to more traditional journals, being Github-based and completely transparent. If you have any questions about the process, please ask!

Top menus are side by side in Python 3.11. Values ​​are only visible when scanned. I also removed the package limitations (== --> >=) and it works fine.
![p2](https://github.com/openjournals/joss-reviews/assets/5471497/e07a9a25-531c-4458-b70a-f7d27c3a1bbb)

![purple1](https://github.com/openjournals/joss-reviews/assets/5471497/bc734e54-93ca-46a6-8713-1e54d550c6ee)
