### 2023-07-21 #5673 @DevenPanchal
Thanks @IceKhan13 for the responses. @lucydot  I have reviewed the entire checklist, and can confirm the installation and Functionality claims for the software are also as per the Paper and the documentation. You have my approval. Thanks!
