+++ api_server/core/migrations/0003_trial_versions.py
@@ -1,0 +1,21 @@
+"""Add versions migration."""
+
+# Generated by Django 4.2.1 on 2023-05-12 20:11
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+    """Migration"""
+
+    dependencies = [
+        ("core", "0001_initial"),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name="trial",
+            name="versions",
+            field=models.TextField(default=[]),
+        ),
+    ]
