+++ api_server/core/migrations/0003_trial_versions.py
@@ -1,7 +1,5 @@
 """Add versions migration."""
 
-# Generated by Django 4.2.1 on 2023-05-12 20:11
-
 from django.db import migrations, models
 
 
