+++ client/tests/test_trial.py
@@ -54,7 +54,9 @@ class TestTrial(TestCase):
         trial.read(trial_id=uuid)
         self.assertTrue(os.path.isfile(os.path.join(self.save_path, f"{uuid}.json")))
         self.assertEqual(trial.metrics, [["test_metric", 42]])
-        self.assertEqual(trial.versions, [[key, value] for key, value in __qiskit_version__.items()])
+        self.assertEqual(
+            trial.versions, [[key, value] for key, value in __qiskit_version__.items()]
+        )
 
     def test_add_trial(self):
         """Test adding stuff into Trial."""
