+++ api_server/core/migrations/0003_trial_versions.py
@@ -1,0 +1,19 @@
+"""Add versions migration."""
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+    """Migration"""
+
+    dependencies = [
+        ("core", "0002_trial_description"),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name="trial",
+            name="versions",
+            field=models.JSONField(default=list),
+        ),
+    ]
