+++ _sources/stubs/purplecaffeine.Trial.rst.txt
@@ -23,6 +23,7 @@
       ~Trial.add_parameter
       ~Trial.add_tag
       ~Trial.add_text
+      ~Trial.add_version
       ~Trial.export_to_shared_file
       ~Trial.import_from_shared_file
       ~Trial.read
