### 2023-07-23 5fdae01
[BUG] Fix widget scroller (#84)
* fix widget scroller
* update requirements
