### 2023-07-23 b4bbe5d
Create dependabot.yml (#86)
