+++ .github/dependabot.yml
@@ -1,0 +1,17 @@
+# To get started with Dependabot version updates, you'll need to specify which
+# package ecosystems to update and where the package manifests are located.
+# Please see the documentation for all configuration options:
+# https://docs.github.com/github/administering-a-repository/configuration-options-for-dependency-updates
+
+version: 2
+updates:
+  - package-ecosystem: "pip"
+    directory: "/client"
+    schedule:
+      interval: "weekly"
+    open-pull-requests-limit: 4
+  - package-ecosystem: "pip"
+    directory: "/api_server"
+    schedule:
+      interval: "weekly"
+    open-pull-requests-limit: 4
