### 2023-07-23 45a81dd
Update ipywidgets requirement from ~=8.0.6 to ~=8.0.7 in /client
Updates the requirements on [ipywidgets](https://github.com/jupyter-widgets/ipywidgets) to permit the latest version.
- [Release notes](https://github.com/jupyter-widgets/ipywidgets/releases)
- [Commits](https://github.com/jupyter-widgets/ipywidgets/compare/8.0.6...8.0.7)

---
updated-dependencies:
- dependency-name: ipywidgets
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
