+++ client/requirements.txt
@@ -2,7 +2,7 @@ qiskit>=0.43.0
 qiskit-ibm-runtime>=0.10.0
 pympler~=0.9
 boto3>=1.26.133
-ipywidgets~=8.0.6
+ipywidgets~=8.0.7
 matplotlib>=3.7.1
 pandas>=2.0.2
 pylatexenc>=2.10
