### 2023-07-24 b3d7a26
Update requirements.txt
Updated requirements for ipywidgets
