### 2023-07-24 cdc34d4
Update pympler requirement from ~=0.9 to ~=1.0 in /client
Updates the requirements on [pympler](https://github.com/pympler/pympler) to permit the latest version.
- [Changelog](https://github.com/pympler/pympler/blob/master/CHANGELOG.md)
- [Commits](https://github.com/pympler/pympler/compare/0.9...1.0.1)

---
updated-dependencies:
- dependency-name: pympler
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
