+++ client/requirements.txt
@@ -1,6 +1,6 @@
 qiskit>=0.43.0
 qiskit-ibm-runtime>=0.10.0
-pympler~=0.9
+pympler~=1.0
 boto3>=1.26.133
 ipywidgets>=8.0.7
 matplotlib>=3.7.1
