+++ api_server/requirements.txt
@@ -6,4 +6,4 @@ djangorestframework-simplejwt==5.2.2
 gunicorn==20.1.0
 django-health-check==3.17.0
 whitenoise==6.5.0
-drf-spectacular==0.26.2
+drf-spectacular==0.26.4
