+++ api_server/requirements.txt
@@ -3,7 +3,7 @@ djangorestframework==3.14.0
 psycopg2-binary==2.9.6
 tzdata==2023.3
 djangorestframework-simplejwt==5.2.2
-gunicorn==20.1.0
+gunicorn==21.2.0
 django-health-check==3.17.0
 whitenoise==6.5.0
 drf-spectacular==0.26.4
