### 2023-07-24 3e8ecaa
Bump django from 4.2.1 to 4.2.3 in /api_server (#91)
Bumps [django](https://github.com/django/django) from 4.2.1 to 4.2.3.
- [Commits](https://github.com/django/django/compare/4.2.1...4.2.3)

---
updated-dependencies:
- dependency-name: django
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
