+++ api_server/requirements.txt
@@ -1,4 +1,4 @@
-Django==4.2.1
+Django==4.2.3
 djangorestframework==3.14.0
 psycopg2-binary==2.9.6
 tzdata==2023.3
