### 2023-08-15 #5673 @obliviateandsurrender
Thanks, @IceKhan13! The package looks good in its current state. The documentation and code quality appear to be great. I just have some small comments before I can give approval from my end:

1. `Guides` tiles have no thumbnails in the documentation. It would be good to have them.
2. Pip installation doesn't work when done using `pip install .`, since the version number is 0.0.0 in the repository.
3. There has been no mention of what kind of serialization is used in storing experimental data. From the code, I could figure out the authors are using `JSON` with `pickle`. A discussion on this would be helpful both for the paper and documentation. Also, it is critical to add a warning about using `pickle` not being secured and unpickling only the data that they trust.
4. The example trial hashes are long and weird. I do understand they are required for uniquely determining an experiment result. But maybe from a UI perspective, it would be better to use something else (maybe normal indexing) for the left side of the widget setups.
5. It might be good to have unit testing for the APIs - basically the `Django` implementation. This seems to be missing right now.
