### 2023-08-21 #5673 @lucydot
Thank-you @obliviateandsurrender for your points raised, @IceKhan13 over to you!

RE: tests. JOSS does not enforce a particular code coverage for unit tests. The documentation states:

> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

If the code is deemed fragile due to lacking tests then we may need to think about how improvements can be implemented.
