### 2023-08-22 IceKhan13/purplecaffeine#99 @IceKhan13
API service: add tests
### What is the expected enhancement?

Add endpoint tests for API service.

Ref: https://www.django-rest-framework.org/api-guide/testing/#api-test-cases

Required to address JoSS review comments https://github.com/openjournals/joss-reviews/issues/5673#issuecomment-1679076266
