+++ docs/index.rst
@@ -68,6 +68,20 @@ Step 2: visualize tracked data
 
 ------------
 
+**Serialization**
+
+Trials are serialized using Python `json` module. We've implemented
+custom `json.JSONEncoder` and `json.JSONDecoder` to handle quantum computing objects
+like `QuantumCircuit`, `Backend`, `Operator`, etc.
+
+Encoders and decoders are using `pickle` module to serialize some classes (`Backend`).
+The pickle module is not secure. Only unpickle data you trust.
+It is possible to construct malicious pickle data which will
+execute arbitrary code during unpickling. Never unpickle data that
+could have come from an untrusted source, or that could have been tampered with.
+
+------------
+
 **Content**
 
 .. toctree::
