### 2023-08-25 9224633
autom django tests
