+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 minversion = 3.8
-envlist = py38, py39, py310, lint, coverage, black, ecosystem, docs
+envlist = py38, py39, py310, lint, coverage, django, lint, black, ecosystem, docs, jupyter
 skipsdist = True
 
 [testenv]
@@ -14,14 +14,20 @@ setenv =
 deps = -r {env:FOLDER:}/requirements.txt
        -r requirements-dev.txt
 commands =
-  pip check {env:FOLDER:}
-  python -m unittest discover -s {env:FOLDER:} -v
+  pip check client
+  python -m unittest discover -s client -v
 
 [testenv:coverage]
 basepython = python3
 commands =
-  coverage3 run -m unittest discover -s {env:FOLDER:} -v
+  coverage3 run -m unittest discover -s client -v
   coverage3 report --fail-under=80
+
+[testenv:django]
+basepython = python3
+commands =
+  pip check api_server
+  python3 api_server/manager.py test
 
 [testenv:lint]
 envdir = .tox/lint
