### 2023-08-25 8749cc7
typo + var is better than hard even for 1 possibility
