+++ tox.ini
@@ -14,20 +14,20 @@ setenv =
 deps = -r {env:FOLDER:}/requirements.txt
        -r requirements-dev.txt
 commands =
-  pip check client
-  python -m unittest discover -s client -v
+  pip check {env:FOLDER:}
+  python -m unittest discover -s {env:FOLDER:} -v
 
 [testenv:coverage]
 basepython = python3
 commands =
-  coverage3 run -m unittest discover -s client -v
+  coverage3 run -m unittest discover -s {env:FOLDER:} -v
   coverage3 report --fail-under=80
 
 [testenv:django]
 basepython = python3
 commands =
-  pip check api_server
-  python3 api_server/manager.py test
+  pip check {env:FOLDER:}
+  python3 {env:FOLDER:}/manage.py test {env:FOLDER:}/
 
 [testenv:lint]
 envdir = .tox/lint
