### 2023-08-25 b1fc6b8
add asset test on return data
