### 2023-08-25 c640318
add coverage into django test
