+++ api_server/requirements.txt
@@ -7,3 +7,4 @@ gunicorn==21.2.0
 django-health-check==3.17.0
 whitenoise==6.5.0
 drf-spectacular==0.26.4
+django-nose==1.4.7
