+++ tox.ini
@@ -27,7 +27,8 @@ commands =
 basepython = python3
 commands =
   pip check {env:FOLDER:}
-  python3 {env:FOLDER:}/manage.py test {env:FOLDER:}/
+  coverage3 {env:FOLDER:}/manage.py test {env:FOLDER:}/
+  coverage3 report --fail-under=80
 
 [testenv:lint]
 envdir = .tox/lint
