+++ tox.ini
@@ -27,7 +27,7 @@ commands =
 basepython = python3
 commands =
   pip check {env:FOLDER:}
-  coverage3 {env:FOLDER:}/manage.py test {env:FOLDER:}/
+  coverage3 run {env:FOLDER:}/manage.py test {env:FOLDER:}/
   coverage3 report --fail-under=80
 
 [testenv:lint]
