### 2023-08-25 c9942c8
API: Add unit tests (#100)
