+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 minversion = 3.8
-envlist = py38, py39, py310, lint, coverage, black, ecosystem, docs
+envlist = py38, py39, py310, lint, coverage, django, lint, black, ecosystem, docs, jupyter
 skipsdist = True
 
 [testenv]
@@ -21,6 +21,13 @@ commands =
 basepython = python3
 commands =
   coverage3 run -m unittest discover -s {env:FOLDER:} -v
+  coverage3 report --fail-under=80
+
+[testenv:django]
+basepython = python3
+commands =
+  pip check {env:FOLDER:}
+  coverage3 run {env:FOLDER:}/manage.py test {env:FOLDER:}/
   coverage3 report --fail-under=80
 
 [testenv:lint]
