+++ _sources/guides/04_setup_api.ipynb.txt
@@ -209,7 +209,7 @@
     "        \"metrics\": [[\"nb_qubits\", 2]],\n",
     "        \"parameters\": [[\"OS\", \"ubuntu\"]],\n",
     "        \"circuits\": [],\n",
-    "        \"operators\": [[\"obs\", Pauli(\"XZYI\")]],\n",
+    "        \"operators\": [],\n",
     "        \"artifacts\": [],\n",
     "        \"texts\": [],\n",
     "        \"arrays\": [],\n",
@@ -236,7 +236,7 @@
     "        \"metrics\": [[\"nb_qubits\", 2]],\n",
     "        \"parameters\": [[\"OS\", \"ubuntu\"]],\n",
     "        \"circuits\": [],\n",
-    "        \"operators\": [[\"obs\", Pauli(\"XZYI\")]],\n",
+    "        \"operators\": [],\n",
     "        \"artifacts\": [],\n",
     "        \"texts\": [],\n",
     "        \"arrays\": [],\n",
