### 2023-08-25 #5673 @IceKhan13
Hello-hello!

We've addressed all comments. In more details:

> Guides tiles have no thumbnails in the documentation. It would be good to have them.

Added default thumbnails if there are no images within notebook to use.
Addressed in https://github.com/IceKhan13/purplecaffeine/pull/98

> Pip installation doesn't work when done using pip install ., since the version number is 0.0.0 in the repository.

Client itself is located under `client` directory. `pip install .` works there.

> There has been no mention of what kind of serialization is used in storing experimental data. From the code, I could figure out the authors are using JSON with pickle. A discussion on this would be helpful both for the paper and documentation.
 > Also, it is critical to add a warning about using pickle not being secured and unpickling only the data that they trust.
The example trial hashes are long and weird. I do understand they are required for uniquely determining an experiment result. But maybe from a UI perspective, it would be better to use something else (maybe normal indexing) for the left side of the widget setups.

Added to docs and paper serialization section.
Addressed in https://github.com/IceKhan13/purplecaffeine/pull/98

> It might be good to have unit testing for the APIs - basically the Django implementation. This seems to be missing right now.

API tests has been added.
Addressed in https://github.com/IceKhan13/purplecaffeine/pull/100

@obliviateandsurrender thank you for great review and precise comments!

@lucydot we hope you will have a great time off! :)

Let us know if need to add anything else. Back to you!
