### 2023-08-28 1f022d3
Bump djangorestframework-simplejwt from 5.2.2 to 5.3.0 in /api_server
Bumps [djangorestframework-simplejwt](https://github.com/jazzband/djangorestframework-simplejwt) from 5.2.2 to 5.3.0.
- [Release notes](https://github.com/jazzband/djangorestframework-simplejwt/releases)
- [Changelog](https://github.com/jazzband/djangorestframework-simplejwt/blob/master/CHANGELOG.md)
- [Commits](https://github.com/jazzband/djangorestframework-simplejwt/compare/v5.2.2...v5.3.0)

---
updated-dependencies:
- dependency-name: djangorestframework-simplejwt
  dependency-type: direct:production
  update-type: versi
[…]
