+++ api_server/requirements.txt
@@ -2,7 +2,7 @@ Django==4.2.4
 djangorestframework==3.14.0
 psycopg2-binary==2.9.7
 tzdata==2023.3
-djangorestframework-simplejwt==5.2.2
+djangorestframework-simplejwt==5.3.0
 gunicorn==21.2.0
 django-health-check==3.17.0
 whitenoise==6.5.0
