### 2023-08-31 #5673 @obliviateandsurrender
Thank you for the quick updates @IceKhan13! One final minor nitpick regarding installation: it would be good to mention in the readme regarding installation via source, i.e., one needs to be in the client folder to do so.
Hey @lucydot, I have updated the checkboxes. It is good to go from my end once the above is addressed.
