### 2023-09-01 fd436ee
Added installation instructions for client
