### 2023-09-03 1e4797d
Update README.md
Added additional install instructions to quickstart section
