+++ client/README.md
@@ -30,6 +30,11 @@ This project is aimed to create simple general interface to track quantum experi
 
 ### Quickstart
 
+```shell
+pip install purplecaffeine
+```
+
+
 ```python
 from qiskit.circuit.random import random_circuit
 from qiskit.quantum_info.random import random_pauli
@@ -76,7 +81,9 @@ Widget(local_storage).show()
 If you want to install from source code, execute while in the root folder of the project:
 
 ```shell
-pip install client/.
+cd ./client
+pip install .
+```
 
 ----------------------------------------------------------------------------------------------------
 
