### 2023-09-03 065ff8f
Added installation instructions for client (#104)
* Added installation instructions for client

---------

Co-authored-by: Michaël <mika89700@gmail.com>
Co-authored-by: Iskandar Sitdikov <IceKhan13@users.noreply.github.com>
