### 2023-09-05 e8519ac
small changes to wording
Please check the changes to ensure that your meaning is not lost
