### 2023-09-05 #5673 @lucydot
Hi @IceKhan13 I've made some small changes to the wording of your paper at https://github.com/IceKhan13/purplecaffeine/pull/105
