### 2023-09-05 #5673 @lucydot
@IceKhan13 can you confirm that the arxiv publication from Cross et al does not have a peer reviewed version that may be more suitable as a reference (e.g., https://dl.acm.org/doi/full/10.1145/3505636)
