### 2023-09-05 cdbdd79
Remove authors field in QAMP entry (#106)
This is to attempt to remove the floating "," in the compiled paper. Will need to attempt compiling to see if successful! If not, then may need to add an authors field with a suitable entry (ie, not leave it empty as this creates the floating comma).
