+++ docs/paper/paper.bib
@@ -6,7 +6,6 @@
 }
 
 @misc{qamp,
-	author = {},
 	title = {Qiskit Advocate Mentorship Program},
 	year = {2023},
 	publisher = {GitHub},
