### 2023-09-05 #5673 @IceKhan13
> can you confirm that the arxiv publication from Cross et al does not have a peer reviewed version that may be more suitable as a reference (e.g., https://dl.acm.org/doi/full/10.1145/3505636)

Yes, authors even uses this reference to arxiv in per reviewed paper thank you've linked.

> I've made some small changes to the wording of your paper at https://github.com/IceKhan13/purplecaffeine/pull/105
> another PR here: https://github.com/IceKhan13/purplecaffeine/pull/106

Both PRs merged! Thank you!
