### 2023-09-19 c139f3e
Create CITATION.cff
