+++ CITATION.cff
@@ -13,6 +13,6 @@ authors:
   - family-names: Hérnandez
     given-names: Luis Eduardo Martínez
     orcid: https://orcid.org/0000-0003-0354-5092
-title: IceKhan13/purplecaffeine: v0.2.4
+title: PurpleCaffeine: tracking of quantum programs and experiments
 version: v0.2.4
 date-released: 2023-09-19
