### 2023-09-19 a519559
Create CITATION.cff (#109)
