### 2023-09-19 #5673 @IceKhan13
> Double check authors and affiliations (including ORCIDs)

✅

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Version `0.2.5`
https://github.com/IceKhan13/purplecaffeine/releases/tag/v0.2.5

> Archive the release on Zenodo/figshare/etc and post the DOI here.

[DOI: 10.5281/zenodo.8357519](https://doi.org/10.5281/zenodo.8357519)

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

✅

> Make sure that the license listed for the archive is the same as the software license.

✅

@lucydot done!
