+++ api_server/requirements.txt
@@ -6,5 +6,5 @@ djangorestframework-simplejwt==5.3.0
 gunicorn==21.2.0
 django-health-check==3.17.0
 whitenoise==6.5.0
-drf-spectacular==0.26.4
+drf-spectacular==0.26.5
 django-nose==1.4.7
