+++ tox.ini
@@ -11,34 +11,34 @@ setenv =
   VIRTUAL_ENV={envdir}
   LANGUAGE=en_US
   LC_ALL=en_US.utf-8
-deps = -r {env:FOLDER:}/requirements.txt
+deps = -r {env:FOLDER:client}/requirements.txt
        -r requirements-dev.txt
 commands =
   pip check {env:FOLDER:}
-  python -m unittest discover -s {env:FOLDER:} -v
+  python -m unittest discover -s {env:FOLDER:client} -v
 
 [testenv:coverage]
 basepython = python3
 commands =
-  coverage3 run -m unittest discover -s {env:FOLDER:} -v
+  coverage3 run -m unittest discover -s {env:FOLDER:client} -v
   coverage3 report --fail-under=80
 
 [testenv:django]
 basepython = python3
 commands =
   pip check {env:FOLDER:}
-  coverage3 run {env:FOLDER:}/manage.py test {env:FOLDER:}/
+  coverage3 run {env:FOLDER:api_server}/manage.py test {env:FOLDER:api_server}/
   coverage3 report --fail-under=80
 
 [testenv:lint]
 envdir = .tox/lint
 commands =
-  pylint -rn {env:FOLDER:}
+  pylint -rn {env:FOLDER:client}
 
 [testenv:black]
 envdir = .tox/lint
 deps = -r requirements-dev.txt
-commands = black {posargs} {env:FOLDER:} --check
+commands = black {posargs} {env:FOLDER:client} --check
 
 [testenv:ecosystem]
 skip_install = true
