+++ api_server/requirements.txt
@@ -1,6 +1,6 @@
 Django==4.2.5
 djangorestframework==3.14.0
-psycopg2-binary==2.9.7
+psycopg2-binary==2.9.8
 tzdata==2023.3
 djangorestframework-simplejwt==5.3.0
 gunicorn==21.2.0
